requirements.txt
alembic==1.4.2
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.2
beautifulsoup4==4.9.1
cffi==1.14.0
click==7.1.2
colorama==0.4.3
colored==1.4.2
distlib==0.3.0
Django==3.0.6
django-bootstrap4==2.0.1
dnspython==1.16.0
dominate==2.5.1
filelock==3.0.12
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.1.2
prettytable==0.7.2
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pylint==2.4.4
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pyperclip==1.8.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
six==1.14.0
soupsieve==2.0.1
SQLAlchemy==1.3.16
sqlparse==0.3.1
virtualenv==20.0.18
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.11.2
